requirements.txt


langgraph
langchain-google-genai
google-generativeai
pydantic
python-dotenv



PyQt6

playwright



SpeechRecognition
pyttsx3
PyAudio



pillow
requests